Add mobile version to user page
