+++ styles/Header/Header.module.css
@@ -84,5 +84,6 @@
   word-wrap: break-word;
   width: 40%;
   font-size: 1.75rem;
+  line-height: 2.5rem;
 }
 /*# sourceMappingURL=Header.module.css.map */
