+++ styles/Header/Header.module.scss
@@ -20,10 +20,6 @@
       width: 30%;
       height: fit-content;
       display: block;
-
-      // .image {
-      //   object-fit: cover;
-      // }
     }
     .description {
       color: $textColor;
@@ -59,6 +55,7 @@
       word-wrap: break-word;
       width: 40%;
       font-size: 1.75rem;
+      line-height: 2.5rem;
     }
   }
 }
