+++ styles/Nav/Desktop/Nav.module.scss
@@ -1,0 +1,44 @@
+@import "../../variables";
+
+.nav {
+  width: 100vw;
+  height: $navHeight;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: $mainColor;
+
+  .logo {
+    width: 120px;
+    height: 120px;
+    display: block;
+    margin: auto 30px;
+
+    // .image {
+    //   object-fit: cover;
+    // }
+  }
+
+  .navigation {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: auto 30px;
+
+    // .underlined {
+    //   text-decoration: underline;
+    //   text-decoration-color: $borderColor;
+    // }
+
+    a {
+      color: $textColor;
+      font-size: 2rem;
+      margin: auto 20px;
+      display: flex;
+
+      &:hover {
+        color: $colorOne;
+      }
+    }
+  }
+}
