+++ styles/User/Desktop/User.module.scss
@@ -1,0 +1,51 @@
+@import "../../variables";
+
+.user_details {
+  width: fit-content;
+  height: fit-content;
+  color: $textColor;
+
+  .detail_item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin: 10px;
+    .detail_item_text {
+      font-size: 1.5rem;
+      font-weight: bold;
+      margin-right: 20px;
+
+      span {
+        vertical-align: middle;
+      }
+    }
+  }
+}
+.edit_box {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: $colorTwo;
+  position: absolute;
+  width: 100vw;
+  height: 100vh;
+  top: $navHeight;
+
+  p {
+    padding: 10px;
+  }
+
+  form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+  }
+
+  .two_button_container {
+    display: flex;
+    width: 250px;
+    justify-content: space-between;
+  }
+}
