+++ styles/_variables.scss
@@ -15,6 +15,7 @@ $processingColor: #d3aa14;
 
 //* NavBar height
 $navHeight: 130px;
+$navHeightMobile: 100px;
 
 //* Sass variables that define breakpoint values
 $breakpoints: (
